+++ src/singletons/Settings.hpp
@@ -346,6 +346,12 @@ public:
     BoolSetting lowercaseDomains = {"/links/linkLowercase", true};
 
     /// Streamer Mode
+    // TODO: Should these settings be converted to booleans that live outside of
+    // streamer mode?
+    // Something like:
+    //  - "Hide when streamer mode is enabled"
+    //  - "Always hide"
+    //  - "Don't hide"
     EnumSetting<StreamerModeSetting> enableStreamerMode = {
         "/streamerMode/enabled", StreamerModeSetting::DetectStreamingSoftware};
     BoolSetting streamerModeHideUsercardAvatars = {
@@ -356,6 +362,10 @@ public:
         "/streamerMode/hideViewerCountAndDuration", false};
     BoolSetting streamerModeHideModActions = {"/streamerMode/hideModActions",
                                               true};
+    BoolSetting streamerModeHideSuspiciousUsers = {
+        "/streamerMode/hideSuspiciousUsers",
+        true,
+    };
     BoolSetting streamerModeMuteMentions = {"/streamerMode/muteMentions", true};
     BoolSetting streamerModeSuppressLiveNotifications = {
         "/streamerMode/supressLiveNotifications", false};