+++ src/singletons/StreamerMode.cpp
@@ -207,6 +207,11 @@ bool StreamerMode::shouldHideModActions() const
     return getSettings()->streamerModeHideModActions && this->isEnabled();
 }
 
+bool StreamerMode::shouldHideSuspiciousUsers() const
+{
+    return getSettings()->streamerModeHideSuspiciousUsers && this->isEnabled();
+}
+
 void StreamerMode::start()
 {
     this->private_->start();