+++ src/singletons/StreamerMode.hpp
@@ -23,6 +23,9 @@ public:
     /// Returns true if streamer mode is enabled & the settings to hide mod actions is enabled
     [[nodiscard]] virtual bool shouldHideModActions() const = 0;
 
+    /// Returns true if streamer mode is enabled & the settings to hide messages from suspicious users is enabled
+    [[nodiscard]] virtual bool shouldHideSuspiciousUsers() const = 0;
+
     virtual void start() = 0;
 
 Q_SIGNALS:
@@ -43,6 +46,7 @@ public:
     bool isEnabled() const override;
 
     bool shouldHideModActions() const override;
+    bool shouldHideSuspiciousUsers() const override;
 
     void start() override;